Remove excessive files.
