--- javadocs-checker/plugin/src/main/java/org/spine3/gradle/javadoc/JavadocCheckerPlugin.java
@@ -1,53 +1,0 @@
-/*
- * Copyright 2016, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.spine3.gradle.javadoc;
-
-import org.gradle.api.Plugin;
-import org.gradle.api.Project;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.spine3.gradle.javadoc.fqnchecker.FqnCheckPlugin;
-
-/**
- * Plugin to check javadocs.
- *
- * @author Alexander Aleksandrov
- */
-public class JavadocCheckerPlugin implements Plugin<Project> {
-
-    private static final String SPINE_JAVADOC_EXTENSION_NAME = "spineJavadoc";
-
-    @Override
-    public void apply(Project project) {
-        project.getExtensions().create(SPINE_JAVADOC_EXTENSION_NAME, Extension.class);
-        log().debug("Applying Spine javadocChecker plugin");
-        new FqnCheckPlugin().apply(project);
-    }
-
-    private static Logger log() {
-        return LoggerSingleton.INSTANCE.logger;
-    }
-
-    private enum LoggerSingleton {
-        INSTANCE;
-        @SuppressWarnings("NonSerializableFieldInSerializableClass")
-        private final Logger logger = LoggerFactory.getLogger(JavadocCheckerPlugin.class);
-    }
-}
