+++ javadocs-checker/plugin/src/main/resources/META-INF/gradle-plugins/org.spine3.tools.javadocs-checker.properties
@@ -19,4 +19,4 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 
-implementation-class=org.spine3.gradle.javadoc.JavadocCheckerPlugin
+implementation-class=org.spine3.gradle.javadoc.fqnchecker.FqnCheckPlugin
